feat(attention): impl self attention

+++ src/little_language_model/tokenizer.py
@@ -15,8 +15,8 @@ class GPTDataset(Dataset):
         for i in range(0, len(token_ids) - max_length, stride):
             input_chunk = token_ids[i:i + max_length]
             target_chunk = token_ids[i + 1:i + max_length + 1]
-            self.input_ids.append(torch.tensor(input_chunk))
-            self.target_ids.append(torch.tensor(target_chunk))
+            self.input_ids.append(torch.tensor(input_chunk, dtype=torch.float))
+            self.target_ids.append(torch.tensor(target_chunk, dtype=torch.float))
 
     def __len__(self):
         """
